removed the please share and added my own things.
I added what I did today, as I am going to e trying to code the robot Mr. Hossain gave me today in order to learn a little more about the coding of it. and I also did try learning some java out of the fist slides that I saw and understood some of the expressions and logical functions.